Revert "Remove platform requirements (#37)"
This reverts commit fa2eb2572b3dad344d15c0411767a4415f347626.
