+++ .github/workflows/ci.yml
@@ -16,6 +16,8 @@ jobs:
     strategy:
       matrix:
         xcode:
+          - '12.4'
+          - '12.5.1'
           - '13.0'
           - '13.1'
           - '13.2.1'
