+++ Package.swift
@@ -4,6 +4,12 @@ import PackageDescription
 
 let package = Package(
   name: "swift-custom-dump",
+  platforms: [
+    .iOS(.v13),
+    .macOS(.v10_15),
+    .tvOS(.v13),
+    .watchOS(.v6),
+  ],
   products: [
     .library(
       name: "CustomDump",
