--- Sources/CustomDump/Internal/CollectionDifference.Change.swift
@@ -1,8 +1,0 @@
-extension CollectionDifference.Change {
-  var offset: Int {
-    switch self {
-    case let .insert(offset, _, _), let .remove(offset, _, _):
-      return offset
-    }
-  }
-}
