+++ Tests/CustomDumpTests/Conformances/UserNotificationsTests.swift
@@ -5,7 +5,6 @@
 
   class UserNotificationsTests: XCTestCase {
     func testUNAuthorizationOptions() {
-      guard #available(iOS 10, tvOS 10, *) else { return }
       var dump: String = ""
       customDump([.badge, .alert] as UNAuthorizationOptions, to: &dump)
       XCTAssertEqual(
